+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Add parameter support for test commands
 * Cmdt no longer breaks when YAML file is empty
 * Fix test command execution error propagation
+* Change min node engine to >= 0.10.0
 
 ### 0.0.5
 * Fix sample test file on README.